:sparkles: Commonize ClickableLinkText. Use ClickableLinkText, which has been made into a common component, on the About screen.
